+++ src/Traits/ElementSaveRelation.php
@@ -67,6 +67,10 @@ trait ElementSaveRelation
             $values = $this->getValueFromModel();
         }
 
+        //for model hasmany->multiselect
+        if(is_null($values));
+            $values = $request->input($this->getPath());
+
         $relation = $this->getModel()->{$attribute}();
         if ($relation instanceof \Illuminate\Database\Eloquent\Relations\BelongsToMany) {
             $this->syncBelongsToManyRelation($relation, $values);